gap/requirements.txt
Django
psycopg2
channels
Pillow
Redis
channels-redis
djangorestframework
django-filter